修改了一下代码,健身计数测试通过,健身规划按实际情况进行了修改

+++ Fitness-Equipment/application/UnitreeMotorSerialService/unitreemotorserialservice.h
@@ -29,7 +29,7 @@ signals:
     void sendUnitreeMotorDataToEquipmentConnectWindow(UnitreeReceive receivemessage);
 public slots:
     //设置力矩
-    void setMoment(int moment);
+    void setMoment(float moment);
 
 private:
     SerialPort *motorSerialPort;
